Handeling exceptions

+++ src/Controller/Admin/Media/MediaDeleteController.php
@@ -4,6 +4,9 @@ namespace App\Controller\Admin\Media;
 
 use App\Entity\Media;
 use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Bundle\SecurityBundle\Security;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpKernel\Attribute\AsController;
 use Symfony\Component\Routing\Annotation\Route;
@@ -14,7 +17,8 @@ final class MediaDeleteController
 {
     public function __construct(
         private EntityManagerInterface $entityManager,
-        private RouterInterface $router
+        private RouterInterface $router,
+        private Security $security
     )
     {}
 
@@ -22,6 +26,17 @@ final class MediaDeleteController
     public function __invoke(int $id): RedirectResponse
     {
         $media = $this->entityManager->getRepository(Media::class)->find($id);
+
+        if (!$media) {
+            throw new NotFoundHttpException("Le média que vous essayez de supprimer n'existe pas.");
+        }
+
+        $user = $this->security->getUser();
+
+        if ($media->getUser() != $user && !$this->security->isGranted('ROLE_ADMIN')) {
+            throw new AccessDeniedException("Vous n'avez pas les droits pour supprimer ce média");
+        }
+
         $this->entityManager->remove($media);
         $this->entityManager->flush();
         unlink($media->getPath());
